+++ generators/server/templates/src/boot/model.js
@@ -1,8 +1,8 @@
 import knex from "knex";
-import { Model } from "objection";
+import objection from "objection";
 
 import knexConfig from "../../knexfile.cjs";
 
 const knexConnection = knex(knexConfig);
 
-Model.knex(knexConnection);
+objection.Model.knex(knexConnection);