+++ generators/server/templates/src/db/migrations/migration.stub.cjs
@@ -1,0 +1,13 @@
+/**
+ * @typedef {import("knex")} Knex
+ */
+
+/**
+ * @param {Knex} knex
+ */
+exports.up = async (knex) => {};
+
+/**
+ * @param {Knex} knex
+ */
+exports.down = function (knex) {};